PR-A5d 拆分验证层编码与算法入口 [skip-slow]
