+++ pyfcstm/verify/__init__.py
@@ -6,6 +6,7 @@ from .inspect_adapter import (
     iter_inspect_eligible,
 )
 from .registry import REGISTRY
+from .result import AlgorithmResult, ResultKind
 from .taxonomy import (
     CALL_COUNT_SCALING_ORDER,
     COMPLEXITY_TIER_ORDER,
@@ -23,6 +24,7 @@ __all__ = [
     "CALL_COUNT_SCALING_ORDER",
     "COMPLEXITY_TIER_ORDER",
     "REGISTRY",
+    "AlgorithmResult",
     "CallCountScaling",
     "Closedness",
     "ComplexityTier",
@@ -30,6 +32,7 @@ __all__ = [
     "FormulaSizeScaling",
     "InspectAccessForbiddenError",
     "SMTLogic",
+    "ResultKind",
     "VerificationScope",
     "VerifyAlgorithmMeta",
     "eligible_for_inspect",
