+++ pyfcstm/verify/encoding/__init__.py
@@ -1,0 +1,10 @@
+"""Verification encoding helpers for FCSTM runtime semantics."""
+
+__all__ = [
+    "expr",
+    "guard",
+    "initial",
+    "lifecycle",
+    "operation",
+    "trigger",
+]
