+++ pyfcstm/verify/encoding/expr.py
@@ -1,0 +1,31 @@
+"""Expression definedness encoding helpers."""
+
+from ._core import (
+    _ConditionPoint,
+    _append_expr_domain_constraints,
+    _binary_z3_or_result,
+    _build_type_constraints,
+    _definedness_feasibility_or_result,
+    _expr_conditions_and_z3_or_result,
+    _expr_to_z3_or_result,
+    _expr_z3_and_domains_or_result,
+    _path_reachability_or_result,
+    _ufunc_z3_or_result,
+    _unary_z3_or_result,
+    _z3_vars,
+)
+
+__all__ = [
+    "_ConditionPoint",
+    "_append_expr_domain_constraints",
+    "_binary_z3_or_result",
+    "_build_type_constraints",
+    "_definedness_feasibility_or_result",
+    "_expr_conditions_and_z3_or_result",
+    "_expr_to_z3_or_result",
+    "_expr_z3_and_domains_or_result",
+    "_path_reachability_or_result",
+    "_ufunc_z3_or_result",
+    "_unary_z3_or_result",
+    "_z3_vars",
+]
