+++ pyfcstm/verify/encoding/initial.py
@@ -1,0 +1,23 @@
+"""Root-initial-path and initializer encoding helpers."""
+
+from ._core import (
+    _InitialPathContext,
+    _InitialPathSearchContext,
+    _build_init_constraints_or_result,
+    _is_root_initial_leaf,
+    _root_initial_path_context,
+    _root_initial_path_contexts,
+    _root_initial_path_transitions,
+    _state_path_from_root,
+)
+
+__all__ = [
+    "_InitialPathContext",
+    "_InitialPathSearchContext",
+    "_build_init_constraints_or_result",
+    "_is_root_initial_leaf",
+    "_root_initial_path_context",
+    "_root_initial_path_contexts",
+    "_root_initial_path_transitions",
+    "_state_path_from_root",
+]
