+++ pyfcstm/verify/encoding/lifecycle.py
@@ -1,0 +1,17 @@
+"""Lifecycle first-cycle encoding helpers."""
+
+from ._core import (
+    _action_operations,
+    _concrete_during_operations,
+    _conditional_conditions_from_expr,
+    _conditional_conditions_from_operations,
+    _enter_condition_descriptors_for_context,
+)
+
+__all__ = [
+    "_action_operations",
+    "_concrete_during_operations",
+    "_conditional_conditions_from_expr",
+    "_conditional_conditions_from_operations",
+    "_enter_condition_descriptors_for_context",
+]
